Clean up transactional duplicate entry point
