+++ contracts/transactional/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "transactional"
+version = "0.1.0"
+edition = "2021"
+publish = false
+
+[dependencies]
+soroban-sdk = { workspace = true }
+
+[dev-dependencies]
+soroban-sdk = { workspace = true, features = ["testutils"] }
